Added setting light state to light, added test, performance maybe bad due to 3 db calls

+++ xlightserver/src/main/java/de/pschiessle/xlight/xlightserver/components/MtsValue.java
@@ -20,7 +20,7 @@ public class MtsValue extends BaseEntity {
 
   private Long valueId;
 
-  private List<BigDecimal> values;
+  private List<Double> values;
 
   @Override
   public boolean equals(Object o) {
